requirements.txt
absl-py
arviz
astropy
cloudpickle
corner
cosmopower-jax
Cython
diffrax==0.4.1
dynesty
emcee
equinox
jax==0.4.28
jaxlib==0.4.28
jaxopt
jaxtyping
joblib
matplotlib
numpy
numpyro
optax
palettable
pathspec
pathtools
pickleshare
pytest
scipy
tensorflow
tqdm
